+++ vsphere/data_source_vsphere_virtual_machine.go
@@ -184,7 +184,7 @@ func dataSourceVSphereVirtualMachineRead(d *schema.ResourceData, meta interface{
 	}
 
 	// Read general VM config info
-	if err := flattenVirtualMachineConfigInfo(d, props.Config); err != nil {
+	if err := flattenVirtualMachineConfigInfo(d, props.Config, client); err != nil {
 		return fmt.Errorf("error reading virtual machine configuration: %s", err)
 	}